Get rid of specific request fragments
