+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/walletconnect/request/WC2RequestFragment.kt
@@ -4,22 +4,28 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.compose.runtime.Composable
 import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.platform.ViewCompositionStrategy
 import androidx.core.os.bundleOf
+import androidx.fragment.app.viewModels
 import androidx.lifecycle.viewmodel.compose.viewModel
-import androidx.navigation.NavController
+import androidx.navigation.findNavController
+import androidx.navigation.navGraphViewModels
 import io.horizontalsystems.bankwallet.R
+import io.horizontalsystems.bankwallet.core.AppLogger
 import io.horizontalsystems.bankwallet.core.BaseFragment
-import io.horizontalsystems.bankwallet.core.slideFromBottom
-import io.horizontalsystems.bankwallet.modules.walletconnect.request.sendtransaction.v2.WC2SendEthereumTransactionRequestFragment
-import io.horizontalsystems.bankwallet.modules.walletconnect.request.signmessage.v2.WC2SignMessageRequestFragment
+import io.horizontalsystems.bankwallet.modules.evmfee.EvmFeeCellViewModel
+import io.horizontalsystems.bankwallet.modules.sendevmtransaction.SendEvmTransactionViewModel
+import io.horizontalsystems.bankwallet.modules.walletconnect.request.sendtransaction.WCRequestModule
+import io.horizontalsystems.bankwallet.modules.walletconnect.request.sendtransaction.WCSendEthereumTransactionRequestViewModel
+import io.horizontalsystems.bankwallet.modules.walletconnect.request.sendtransaction.ui.SendEthRequestScreen
+import io.horizontalsystems.bankwallet.modules.walletconnect.request.signmessage.v2.WC2SignMessageRequestScreen
 import io.horizontalsystems.bankwallet.modules.walletconnect.version2.WC2SendEthereumTransactionRequest
 import io.horizontalsystems.bankwallet.modules.walletconnect.version2.WC2SignMessageRequest
-import io.horizontalsystems.core.findNavController
+import io.horizontalsystems.core.helpers.HudHelper
 
 class WC2RequestFragment : BaseFragment() {
+    private val logger = AppLogger("wallet-connect v2")
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -32,8 +38,42 @@ class WC2RequestFragment : BaseFragment() {
             )
             setContent {
                 val requestId = requireArguments().getLong(REQUEST_ID_KEY)
+                val navController = findNavController()
+                val wc2RequestViewModel = viewModel<WC2RequestViewModel>(factory = WC2RequestViewModel.Factory(requestId))
 
-                WC2RequestScreen(requestId, findNavController())
+                when (val pendingRequest = wc2RequestViewModel.requestData?.pendingRequest) {
+                    is WC2SignMessageRequest -> {
+                        WC2SignMessageRequestScreen(navController, pendingRequest.id)
+                    }
+                    is WC2SendEthereumTransactionRequest -> {
+                        val vmFactory by lazy { WCRequestModule.FactoryV2(pendingRequest.id) }
+                        val viewModel by viewModels<WCSendEthereumTransactionRequestViewModel> { vmFactory }
+                        val sendEvmTransactionViewModel by viewModels<SendEvmTransactionViewModel> { vmFactory }
+                        val feeViewModel by navGraphViewModels<EvmFeeCellViewModel>(R.id.wc2RequestFragment) { vmFactory }
+
+                        sendEvmTransactionViewModel.sendSuccessLiveData.observe(viewLifecycleOwner) { transactionHash ->
+                            viewModel.approve(transactionHash)
+                            HudHelper.showSuccessMessage(
+                                requireActivity().findViewById(android.R.id.content),
+                                R.string.Hud_Text_Done
+                            )
+                            navController.popBackStack()
+                        }
+
+                        sendEvmTransactionViewModel.sendFailedLiveData.observe(viewLifecycleOwner) {
+                            HudHelper.showErrorMessage(requireActivity().findViewById(android.R.id.content), it)
+                        }
+
+                        SendEthRequestScreen(
+                            navController,
+                            viewModel,
+                            sendEvmTransactionViewModel,
+                            feeViewModel,
+                            logger,
+                            R.id.wc2RequestFragment
+                        ) { navController.popBackStack() }
+                    }
+                }
             }
         }
     }
@@ -43,25 +83,5 @@ class WC2RequestFragment : BaseFragment() {
 
         fun prepareParams(requestId: Long) =
             bundleOf(REQUEST_ID_KEY to requestId)
-    }
-}
-
-@Composable
-fun WC2RequestScreen(requestId: Long, navController: NavController) {
-    val viewModel = viewModel<WC2RequestViewModel>(factory = WC2RequestViewModel.Factory(requestId))
-
-    when (val pendingRequest = viewModel.requestData?.pendingRequest) {
-        is WC2SignMessageRequest -> {
-            navController.slideFromBottom(
-                R.id.wc2SignMessageRequestFragment,
-                WC2SignMessageRequestFragment.prepareParams(pendingRequest.id)
-            )
-        }
-        is WC2SendEthereumTransactionRequest -> {
-            navController.slideFromBottom(
-                R.id.wc2SendEthereumTransactionRequestFragment,
-                WC2SendEthereumTransactionRequestFragment.prepareParams(pendingRequest.id)
-            )
-        }
     }
 }
